Add LiberarRevisionSECAIAOCDL endpoint to Muestreos controller
A new HTTP PUT endpoint `LiberarRevisionSECAIAOCDL` has been added to the `WebAPI.Controllers.v1.Operacion` namespace in the `Muestreos.cs` file. This endpoint accepts a body parameter `muestreos` of type `IEnumerable<long>`, and query parameters `esLiberacion` of type `bool`, `estatusId` of type `int`, and an optional `filter` of type `string`.

The method creates and populates a list of filters if the `filter` query parameter is not empty. If the `muestreos` collection is empty, it fetches paginated muestreos data using the `GetMuestreosPaginados` command and extracts the `MuestreoId` from the result. Finally, it sends a `LiberarRevisionSECAIAOCDLCommand` with the `estatusId` and `muestreos` to the mediator and returns the result.