Add CambiarEstatusAsync method to MuestreoRepository
Added a new method `CambiarEstatusAsync` to the `MuestreoRepository` class. This method updates the status of `Muestreo` entities based on provided `muestreosIds` and an `EstatusMuestreo` enum. It returns `false` if no matching entities are found and throws an `ApplicationException` on error. Also, added `using` directive for `Application.DTOs.EvidenciasMuestreo` and removed `System.Linq` directive from `MuestreoRepository.cs`.